feat: user can toggle void on/off
This is becoming more like a crossword creation tool rather than a
program to automatically generate crosswords. Added click handler to
Grid.tsx so the user can click a Cell and the state of the cell will be
flipped from being shaded to being a white square. Added state to
Grid which is initialized from grid.ts. Added models for CellProps and
CellType which are an interface for the Cell component and a type for
the cell data respectively.