feat: user can toggle background of all cells
Added initializeGrid which takes our grid and sets an id to equal the
index. This saves me having to type out the id number in grid.ts. Added
logic so the symmetrical cell does not have its background changed if
the cell is the central cell - the cell at index 84 in a 169 cell grid.
Essentially, the central cell is symmetrical with itself, and so we
would toggle it to, say, black, and then immediately toggle it to white
and the user would see no change. Added id to each of our CellType
objects in our grid.ts array.