docs: update notes.md with various thoughts
Added Assumptions section near top of notes. Added section on automatic
numbering of our clues. In order to do this, we'll have to determine if
a cell occurs at the 'start' of a clue, and in fact determine what
'start' means. In order to do this, it will be useful to figure out
relative positions of cells to each other. For example, if the cell at
index 0 is not a void, then we might want to check if the cell directly
underneath is not a void, and then we know that the cell at 0 is the
start of a down clue. Similarly, it's useful to have indices for the
cells along the edge. Added a util function to return the indices of
the cells along the right edge of the grid.