feat: initialize each cell with info on cells near
Added top, right, bottom, and left props to our CellType. If the cell
above our current index is a void or our current cell is on the top edge
then we set the top prop on our current cell to false. False indicates
that the cell immediately above our cell is a void. We'll do the same
for right, bottom and left. This logic is added to our initialize
function. When it comes time to number our clues, it will hopefully make
it easier to determine which cells are the starting point of our down
and across clues.