feat: toggling dark/light updates clue numbers
Added a number of functions to utils. The functions getCellAbove
and getCellBelow return the cell above/below our current index. We can
then update props on that cell. Added setClueNumbers that iterates over
our grid and sets the clueNumber prop where appropriate. Added
updateSurroundingCells which updates the top, bottom, left and right
props of touching cells when we toggle a cell from light to dark or vice
versa. Added isLeftEdge, isRightEdge, isTopEdge and isBottomEdge which
returns true or false if a cell is on any of the respective edges. Moved
initializeGrid into utils. Added clueNumber to our grid.ts in data.