feat: create an array of clue objects
Added a Clue class that we use to create clue objects. Each clue object
stores information like length, direction (across or down), indices of
the cells that make up the clue, etc. Added Direction enum that is used
to set and check the direction prop on our Clue instances. Added
getClueIndices which returns an array of indices where each index is the
starting cell for a clue. Added createClues which creates an array of
Clue instances and returns that array.