feat: add function to remove "-" and "'" chars
Added removeChars function that iterates over answer. If the raw value
is a hyphen or an apostrophe, we remove it from the string and set this
shorter value to the word pro. We also recalculate the length for this
shorter word prop. The length will be used to determine whether or not
an answer will fit into the crossword.