feat: add funcs to assist with updating clues
The Clue class's answer prop is now an array instead of a string. It
will hold the answer as an array of characters. Added getAcrossOrDown.
It takes a Clue and an array of clues as args. If a clue is an across
clue, it can't intersect with other across clues, so we return an array
of down clues. Added getCluesThatIntersect, which takes this filtered
array of clues and a clue and returns an object for each clue that
intersects with our clue param.