refactor: update some util functions
Refactored getAcrossOrDown. Split into two functions, getAcrossClues and
getDownClues. These only take one param, the array of Clues. Added
intersection prop to the Clue class. After we call createClues, we can
then store our across clues and down clues in two variables. For each
clue, we call setCluesThatIntersect which sets the intersection prop on
each clue. The intersection prop is an array of objects representing the
clues that share cells with the current clue. The intersection prop will
make it easier when we are trying to automate the generating of answers
when constructing our crossword. If we set a value for an answer in a
clue, we can update all linked clues. When we attempt to generate an
answer for one of these linked clues, we can use the existing pattern
of letters to find candidate clues.