feat: add answers with regard to existing answers
As answers are added, clues that have shared cells are also updated.
Assuming 13 across and 13 down are both 13 cells long, and we have
chosen an answer for 13 across, the answer for 13 down is selected using
the letter that occupies the shared cell. We filter our possible answers
that are 13 letters long AND contain the letter already in the top left
cell. When instances of the clue Class are created, once we know how
long each answer is, we now populate the answer array prop with empty
strings. Formerly, it was just an empty array. Added two functions, one
to sort our array of clues by descending length prop. This ensures we
can populate the longer answers first. The other function creates a
regular expression using the contents of our answer array. If
the answer array contains only empty strings, we are free to put any
answer of an appropriate length in there. If it contains some letters,
because we have entered other clues that intersect with this clue, then
when it comes time to fill this answer in during construction, we have
to filter our possible answers.