fix: fix clue answer not correctly initialized
The issue manifested as shared letters in words being replaced when a
later clue answer was selected. This was because down Clue instances
were missing an empty string from their answer arrays when created. This
only occurred for down clues and so the app mostly appeared to be
working correctly.