feat: add some ability to replace previous clues
For a clue where there is no candidate answer, added some code to
identify the shared letters, identify the clues that those shared
letters belong to. Essentially, adding code to achieve the steps I've
outlined in my NOTES.md under "Backtracking." At this point, we've
identified the new patterns we will match against when we omit letters
from the other clues that intersect with the current clue. We've also
identified the clues in which those shared letters appear.