feat: replace intersecting clues to find curr clue
Added code to generate a regexp for each clue that intersects with our
current clue. The current clue is one for which there is no candidate
answer in our word list due to the combination of letters contained
in our current clue. Therefore, we will find alternatives for our
intersecting answers and for each alternative we will check if we can
now find a candidate clue for our current clue.