feat: more code for answer replacement
Added getWordList which takes an answer length and AllAnswers as args
and returns the value of the prop in AllAnswers that matches the length
of the clue. This is a better way to find the list of words that match
our clue answer's length. Remember, that answers2.ts has props 'three',
'four', 'five', etc, which reflect the length of the words stored under
these props. Added getMatches which takes the current answer, a regular
expression and a list of candidate answers and returns an array that
contains objects of type Answer. At this point, we can return an array
of candidate answers to replace our current answer.