feat: more code for finding replacement answers
Added an optional 'letter' prop to the objects contained in the Clue
class' intersection array prop. Now we have a reference to the letter
in each clue that intersects with the current clue we're trying to find
an answer for. Added code in setClueAnswers to populate the letter prop.
Updated notes to clarify where I am.