feat: more code to replace intersecting clues
Went back through code and added copious comments. Complexity is getting
to be a lot of work. For the intersecting clues, we are setting the
letters that are unshared to empty strings, but we also have to set the
letter that is shared with our current clue to an empty string. I added
code to do this, but am not sure that it is working. Need some strategic
console logs. Updated notes.