fix: ensure correct letters removed from answers
Refactored code that I identified as definitely not working as intended.
For each clue that intersects with our current clue (an rClue), we
identify if the clues that intersect with our rClue are complete
answers. If they are, then we cannot set the common cell with rClue to
equal an empty string - that's the intention. Unshared squares are also
set to an empty string. This gives us a pattern of letters and empty
strings for rClue which can be used to create a regex to select an
alternative answer for our rClue. This is the code I refactored. Let's
hope it works!