refactor: add code to stop execution if no answer
Added endLoop boolean to populateClues. The populateClues function
iterates over each clue and calls setClueAnswers. If setClueAnswers
cannot find a candidate answer for a clue, then it will attempt to
find alternative answers and finally return true, which causes the
populateClues execution to break. Effectively, once we hit a clue with
no candidate answers, no further answers will be added and this will
help with untangling whether my code to find alternative answers is
working as intended or not.