fix: fix letters not removed from answers
Finally managed to fix issue where we hit a clue with no candidate
answers. We get the clues that intersect with that clue. We remove
letters from these intersecting clues, where possible, in order to find
new alternative clues. The issue was that letters that could be removed
(because they intersected with unfilled answers) were not being removed.
This came down to using the yourIndex prop instead of myIndex found in
Clue.intersection. Letters are now displayed in cells (kinda - you have
to save again in the IDE).