feat: automatically replace answsers
Added getLetter, which takes the current clue and the rClue, one of the
clues that can be replaced, and returns the letter and the index of the
shared letter in the current clue and rClue. We can then use that to
filter out candidate answers that contain that letter at the same
position. Updated notes with to do.