feat: get unique candidate answers
Added code to push unique candidate answers into an array and push
used letters into an array. The purpose is to generate a list of
candidate answers such that each answer has a unique letter intersecting
with our current clue. If the A in APPLE is shared with a clue for which
we can't find an answer, there is no point trying ASPEN and OPENS and
OFFAL, since if OPENS yields no possible answers neither will OFFAL
since they both start with O and will both yield no possible answers.