fix: array of unique answers initializes correctly
The variable uniqueAnswers gets populated correctly with alternative
answers for clues that intersect our current clue. Each answer is unique
in the sense that if the old answer was APPLE and the A is shared with
our current clue, we only push alternative answers to uniqueAnswers if
the letter at the first position hasn't been used before. If
uniqueAnswers contains words that start with A, B ... O, then we won't
push OFFAL to it because we already have a word that starts with O.