feat: remove empty cells after crossword creation
When all clue answers that can be set have been set, we have a checkbox
that indicates whether or not unfilled cells in the crossword should be
blanked out or set to voids. If empty clues are removd, the crossword
construction is complete. Otherwise, we might have an option for the
user to manually update clues. Added checkbox and state to Grid.tsx.
Updated NOTES.md regarding this.