feat: remove clue from clue list
Added a removeClue function that takes an array of clues and an index.
The index is the gridState index, i.e. for a 13x13 grid it is a number
from 0 to 168. We find the clue(s) that this index is/are part of and
remove it from the clue list. Returns an updated array of clues.