feat: add remove clues function to populate clues
Added calls to removeClues to our populateClues function. Updated
setClueNumbers so that it sets all clue numbers to an empty string
initially. This fixes an issue where a clue was being removed but an
existing grid clue number was not being set to an empty string. Changed
button and label text to better reflect their functionality. Updated
NOTES.md.