refactor: replace dup code with existing function
Replaced some code at the beginning of setClueAnswers with a call to
getMatches which is doing the same thing as the code being replaced.
I want to update getMatches so it rejects answers that already exist
within the puzzle. Therefore, we need to ensure that getMatches is used
whereever we are filtering potential clue answers.