feat: reset clues to find alternative answers
Added getAllMatches function. The original getMatches function excludes
the current answer, but we want to retain it here as we want to create
an array of letters that might occur at a position in our clue answer so
that we can construct regexp patterns for matching. Added
createUniqueLetterList function that creates an array of unique letters
that can occur at the intersection point between two clues. Updated
the handleResetClue click handler in Grid.tsx to call both of these new
functions. At this point, we successfully create an array of letters
that can occur at the intersection between our first incomplete answer
and the first clue that intersects with that incomplete answer.