feat: generate patterns to fill in missing answers
We sort allUniqueLetters which contains objects with an array of letters
and an index. The letters are the letters that may appear at the index
in our incomplete answers. We sort by index ascending. Next we use our
function from ChatGPT to generate combinations. We iterate over all the
combinations. For each combination, we create an array of empty strings
the same length as the incomplete clue's answer. allUniqueLetters and
each combination contains the same number of elements. The index value
of the first object in allUniqueLetters represents the position at which
our first letter in our combo appears in our incomplete clue's answer.
Therefore, we can plug our letters into our array of empty strings and
create a pattern to match against to find answers for our incomplete
clue.