feat: replace and set incomplete clue answers
Added a setClueAnswer function that takes a clue and candidate answers
and sets the answer. Added logic to handleResetClue to: return if there
are no incomplete answers, create a pattern, convert it to a RegExp and
push to an array called patterns, get a word list, get matches from the
word list using the patterns, and set an answer on the incomplete clue
if one is found.