feat: update answers for intersecting clues
Added two functions. The function updateIntersectingClues takes the clue
whose intersecting clues must be updated and a copy of the clueList
state. It returns a list of intersecting clues where the intersecting
letter has changed. It updates the answer array of the intersecting
clues so that the letter at the intersection equals the letter in our
update answer. The function resetClue takes a Clue as arg, copies the
answer array, and sets all non-intersecting letters back to empty
strings. For each of our intersecting clues, added code to create a
pattern, match against a word list, and set their new answer.