feat: regenerate answers until all clues complete
Added resetAllAnswers function which deletes the letter prop from each
cell in gridState, deletes the letter prop from the objects in the
intersection prop of each Clue, and sets the clue.answer to an array of
empty strings. What we retain (don't reset) is the grid structure, and
so we don't reinitialize the app completely as if we had refreshed.
Added a fillGrid state to Grid.tsx. This is a boolean. If set to true,
after generating clue answers we check to see if there are incomplete
answers. If there are, then we reset the state and regenerate answers
again and again until there are no incomplete answers. Added a checkbox
for the fillGrid state labelled Force Fill Grid. fillGrid and
removeEmpty are mutally exclusive, so checking 'Remove Emtpy Cells' will
cause Force Fill Grid to become unchecked.