refactor: move getClues from Root to Grid
Moved getClues from Root to Grid. The getClues function will need access
to React state. It makes sense, then, that it is located in Grid where
app state is located. Updated getClues so that it constructs an array
of objects that we can pass to our serverless getClues function which
in turn requests our AI to generate clues for our clue answers.