feat: take clues from AI and update React state
Updated the getClues function in Grid.tsx. It now receives the response
from the getClues serverless function and updates the clue property of
each clue in our React state. Added an example_ai_response file that
contains the response that we receive back from the Claude Haiku LLM.
This is being used for further development so I don't have to keep
hitting the API which could potentially be expensive.