feat: click a cell to change color or select
Updated SolveGrid.tsx. Added direction state. This is a boolean that
toggles each time the user clicks a cell. This may not be necessary.
Added handleCellClick function which is passed to each SolveCell that
SolveGrid renders. This handles the logic of 'selecting' or changing
the background color of the associated answer on the grid. Currently,
clicking a cell toggles its background from yellow to white. Added a
boolean 'selected' prop to our CellType with a default value of false.
This boolean controls the color of the background of each cell. Added
an optional handleCellClick prop to the CellProps interface.