feat: clicking a cell selects across + down clues
Added resetSelectedCells to utils.ts. Takes an array of CellType objects
as param, and sets the selected prop on each cell to false. Updated code
regarding selecting entire down/across clues when the user clicks a
cell in the grid. At this point, clicking on a cell that is part of both
an across answer and a down answer will initially change the background
color of all cells in the across answer to yellow. The next click will
reset all cell background colors to white and then set the cells in the
down answer to yellow. Further clicks toggle between down and across.