feat: clicking a cell highlights answers/clues
Updated the code for handleCelClick such that it matches the behaviour
of the Guardian Quick Crossword grid. Added setSelection which takes
a grid (array of CellType) and a clue as params. It iterates over the
indices prop of the clue and sets the 'selected' prop for appropriate
cells to true.