feat: the clues are rendered to the screen
Added renderClues function to SolveGrid.tsx. Takes an array of Clues
and a direction as params, filters the clues by the direction and then
maps over the filtered clues returning a list element for each with
the clue text. Uses the id of the clue as a key. Updated styles for
SolveCell. The cells are now sized relative to VW and box sizing is
border box. Added ClueContainer component. Rendered by the Solve route,
ideally this is where are clue text is rendered. This will involve
lifting the state up to main.tsx and then using some state management or
prop drilling. This is a big refactor and I may just skip it.