styles: update styles for rendered clues
Updated renderClues. It now renders a span for clue number and a div
that contains the clue text and the number of letters, words, whether
there is a hypen etc. Each list item is displayed as flex so I can put
adequate space between the clue number and the clue text.