feat: clue numbers are rendered next to clue text
Added a clueNumber prop to each Clue instance. Updated renderClues: it
now takes a copy of the clueList state and sorts the clues in ascending
clueNumber order. It now renders the clueNumber for each clue. Added a
new function, setClueNumbersOnClues, which takes an array of Clues and
an array of CellTypes as params and sets the clueNumber prop on each
clue. This new function is called at the end of populateClues, before we
set React state.