feat: word length with hyphen/space displayed
Added getWordLength function to utils.tsx. This takes a clue and returns
a string that is rendered to the screen. An example string: '(5-3,4)',
indicating that there is a 5-letter word followed by a hyphen, then a
3-letter word, and finally a 4 letter word: BIRDS-EYE VIEW. Added two
more possible answers: BIRDSEYEVIEW and PIECEOFCAKE, since our answers2
did not contain idiomatic phrases.