feat: user can use tab key to cycle through clues
Added useEffect so I can add an event listener to
SovleGrid. The event listener calls a new function
called handleTabPress. At the moment, when tab is
pressed, a copy is taken of the clueList state and
it is sorted by clueNumber, then it is sorted by
direction. We can then update the selectedClue
state by getting the index of the currently
selected clue and incrementing it or setting it to
zero. Essentially, if the user has an answer
selected and they hit the tab key, the selection
will cycle through the keys in ascending
clueNumber order.