feat: focus first cells when tabbing through clues
Added code to pass a ref to SolveCell. Added code to generate an array
of refs in SolveGrid, each ref is then passed to SolveCell. These refs
are then used to focus the first cell of each answer as the user uses
tab to cycle between answers. Still some TypeScript errors to be fixed
but code is functioning as expected.