feat: a down arrow keypress focuses the cell below
Added handleArrowKeyPress to SolveGrid. If the user presses the down
arrow key and the cell below the currently selected cell is not void,
the focus will switch to the cell below. Updated handleTabPress so that
it returns if the key pressed is not the tab key and it also updates
the selectedCell React state.