feat: user can use arrow keys to navigate answers
When an answer in the crossword is selected, the user can use the
arrow keys to move from cell to cell. As the user uses the arrow keys,
the cell that is selected gains focus and the appropriate clue and
answer are highlighted. Added the util function getCluesFromCell, which
takes a CellType and an array of Clues as params and returns the array
of Clue(s) that that CellType belongs to.