feat: pressing alpha key focusses the next cell
Added handleAlphaKey and keys.ts. The handleAlpha key function returns
if the key pressed is not an alphabetic key. Otherwise it mimics the
behaviour of the Guardian Quick Crossword - the current input's content
is updated with the key pressed and the next cell in the answer (if
there is one) is focussed. There are some issues, and these have been
added to the NOTES.md file.