feat: entering a letter focusses the next cell
Defined a handleInputChange in SolveGrid. This is passed to SolveCell as
a prop, also added to CellProps model. When the value of the input in
SolveCell is changed, handleInputChange is called. Added an answer prop
to each Cell type. When the value of an input changes, the answer prop
is set by handleInputChange. It also updates the state and focusses the
next cell (if there is one). Deleted the keys.ts file as we don't need
it. Deleted the function handleAlphaKey from SolveGrid and rolled its
functionality into handleInputChange. Added value and onChange
attributes to our input inside SolveCell.