feat: delete key functions as expected
Completed the logic of the handleDelete function. If a cell is not
selected, the delete key does nothing. If a cell is selected and there
is a value in the input, pressing the delete key removes the value and
the focus does not switch to the previous cell in the answer. If the
cell has no value in it (""), pressing the delete key moves focus to the
previous cell (if there is one).