feat: clicking on a clue highlights the answer
Added handleClueClick. The user can click on the list item that contains
the clue text/details, and the corresponding answer on the grid is
selected. The clue itself is highlighted, and the first cell of the
answer is focussed. Further, all of the related state is updated, the
selectedCell, the selectedClue, and the selected prop for each Cell in
gridState.