refactor: handle alpha keys with onKeyDown
In order to mimic the Guardian Quick Crossword behavior, specifically
where the content of an input is replaced if the user presses a key,
it will be necessary to set the onChange handler to a no-op and handle
alpha key presses in a separate function, `handleAlpha`, using onKeyDown.
Updated NOTES.md.