fix: fix lint errors highlighted by biome
Fixed all errors in SolveGrid.tsx. These ranged from using index as a
key on a component (this would have been fine since the order won't
change), using not null assertions (!, many of these), and not having an
onKeyDown on an element with an onClick - added onKeyDown with a no-op
since the click functionality is handled with a separate key handler
passed to the SolveCell.tsx component.