fix: data persisting on route change
Data was persisting when switching routes from editor to solver.
Specifically, the gridState was persisting but the clueList was being
reset. I don't know why this was happening. The symptom was that when
switching from the editor route to the solver route and back again, the
gridState.letter persisted in state and was being rendered to the screen.
I updated the initializeGrid function to set the ".letter" prop to be
set to "" on initialization. Renamed getClues.ts to getClues.cts to fix
compiler warning. Fixed biome warning in Main.tsx related to no non null.
This was rather a pointless fix and is a perfect candidate for when you
DO want to use a non-null assertion - i.e. an element with the id "root"
does exist. Deleted the unused ClueContainer component.