feat: add "reset grid & answers" button
Adding this button introduced an error as I was not fully resetting the
state. I reset the grid state completely but I simply reset the answers
for the clues. This was fine if the grid wasn't edited. Editing the grid
obviously affects the number of clues, the length of the answers etc.
Editing the grid, generating answers, resetting the grid and answers and
then attempting to generate more answers was causing the app to hang
indefinitely (probably an infinite loop). Ensuring that the clueList
state gets remade when the grid is reset fixed the issue. Added
isModified state, false by default, set to true when the grid is edited
and set to false when the grid is reset. This is used to control whether
buttons are disabled and the style of disabled buttons. Added inline
button styles in Grid.tsx and added some contains for the checkboxes and
associated labels to improve layout.