feat: add input & button to save created puzzles
Refactored getLocalStorage slightly so I can use the function for both
user-created puzzles (saving finalized creations and work in progress)
and puzzles for the SolveGrid component - retrieving saved puzzles from
local storage and also saving progress when in the middle of solving a
puzzle. Added puzzleName state to Grid.tsx.