fix: getLocalStorage was always returning puzzles
When refactoring getLocalStorage and attempting to remove TypeScript
errors/warnings, I introduced a bug whereby I had independent if
statements that would set the return value. This caused a TS error that
told me that my return value was being accessed before initialization.
I changed the if statements to be: if, if-else, and as a result, the
value being returned was always what was being set in the final else.
Fixed by chaining if-elseif-else.