feat: store progress when editing puzzles
When editing a puzzle in the 'editor' route, progress is saved to local
storage on each update. When we refresh or navigate back to the page,
saved progress is retrieved from local storage. Refactored
setLocalStorage so that it takes a string (key) and an object of items
to set in local storage.