refactor: select onChange avoids stale state
Updated the onChange handler for our select element in SolveGrid so that
we store computed values in variables before updating different parts
of state. This should ensure the values are constant for each of our
setState calls. Updated notes to clarify exactly what I need to happen
in the SolveGrid component regarding how the user uses it and what that
means for state and local storage.