fix: selection and resetting state in SolveGrid
Added a "-" option to the select element. Moved the code for the select
onChange into a function, handleSelectChange. Selecting "-" causes the
state of the app to be reset. Clicking the "Reset Answers" button also
calls the handleSelectChange function. Selecting puzzle names from the
dropdown loads the appropriate clues and grid into state. Added a value
attribute to the select element(?!).