fix: avoid mutating the imported 'grid' data
Updated SolveGrid and Grid to avoid mutating the 'grid' data imported
from data/grid.ts. Renamed the import to 'initialGrid.' Used deep copy
when passing initialGrid as a argument to a function - specifically
when initializing the state with useState and when resetting the grid
using initializeGrid. This fixes a bug where I could set the selected
puzzle to the name "-" in the select in SolveGrid, select a clue causing
its background to change to yellow, select another puzzle, and then
switch back to "-" and the clue would still be selected. This was as a
result of mutating the imported 'grid' which is to be used as an initial
state or to return to a base state in the app. Changed the
"Save Crossword" button in Grid to be of type 'submit.' Forgot the input
and button were wrapped in a form and so hitting enter in the input was
causing a refresh. Prevented the default action in the button onClick
handler.