refactor: getClues now takes a param, clues
Refactored getClues in SolveGrid so it takes an array of objects of type
Clue. Previously it was making a shallow copy of the clueList state, but
now it is called in the 'New Random Puzzle' handler, and so has to use
the clues generated by resetAllAnswers. Removed the button to retrieve
clues from the AI API.