fix: error and data responses from getClues.cjs
The getClues.cjs is a Netlify serverless function. It receives the
request from my client and then makes a request to the OpenAI API,
receives that response and returns it to my client. Since the Netlify
timeout limit on requests is 10 seconds, I occasionally hit this limit
and the serverless function throws an error. Now, I had some confusion
over what was happening regarding the responses from OpenAI, errors
generated by the serverless function, and what I was sending to my
client. Responses received from OpenAI are strings (error or data),
errors thrown by the serverless function are objects. The solution: if
we get a good response from OpenAI, just return it without parsing it
and then stringifying it again - it's already a string. For errors, we
have to check if the error is a string. If it's a string, then send it
to the client as is, if it's not a string, then it is an error produced
by the serverless function and we have to stringify before forwarding to
the client. Updated getClues.cjs with this logic.