fix: add better error handling for serverless reqs
When a Netlify serverless function times out, it sends a response but
it doesn't use the body, and so you can't parse it. Parsing such a
response throws an error that should never appear in production.
Refactored SolveGrid's getClues function to deal with this. If the
response is not OK and the body is unused, we just throw an error. Else,
we parse the body and if response is OK and we have a content field, we
attempt to use the response to update our clueList state. Otherwise we
throw an error which would be caused by the objects in content being
malformed.