fix: resolve error in handleClueClick in SolveGrid
Fix a bug where the selectedCell was being set to the clue number of the
currently selected clue instead of the first cell in that clue's answer.
For example, clicking clue 8 Across would set the selectedClue state to
be the 9th (0 index based) element of gridState instead of the first
cell in 8 Across which is the 26th element of gridState. Remove comments
from getClues.cjs (Netlify serverless function).