feat: add grid validation to the editor
Added validateGrid function to Grid.tsx. Added an 'isValid' prop to each
CellType. Added logic to Cell.tsx to dynamically set the border color of
each cell based on the state of isValid. The validateGrid function sets
the isValid prop to true for all cells when first called. It then will
perform various checks to identify invalid grid states. Cells that
contribute to the grid being invalid will have their isValid prop set to
false and this will be rendered in the UI as a red border. Currently,
validateGrid is looking for clue answers that are less than 3 characters
long, which is not allowed, and setting the isValid prop to false for
any cell that is part of that clue answer.