feat: add function to get all cells in row or col
Added getRowOrColumn that takes an index (being the index of a cell in
the grid), a direction (Direction enum), and the grid. Depending on
whether Direction.DOWN or ACROSS was passed in, it returns an array of
all the cells that are in the same row or column of the grid as the cell
with the index that was passed in. This can be used to determine whether
or not an entire row or column in the grid consists entirely of voids,
which would make the grid invalid.