feat: add code to identify islands of answers
An island of answers means the crossword grid is not valid. An island
can be defined as a clue or group of clues that are not connected to the
rest of the crossword answers. An island such as this is created by a
group of voids that stretch across the entire crossword, including voids
that are diagonally contiguous to other voids. Added getCellUpRight,
getCellUpLeft, getCellDownLeft, and getCellDownRight which return the
cell in the direction indicated by the function name. Added
getContiguousVoids which takes a grid and an index as args and if the
cell indicated by the index is void will return an array of the indices
of all voids that touch the current void cell.