feat: add code to create arrays of connected voids
Worked through some logic relating to creating arrays of indices that
reference connected voids. A series of voids that stretch from one side
of the board to the other will cause an invalid grid. Added some nested
while loops to see if I could iterate over some sample arrays of numbers
to solve the problem. Determined not to ask ChatGPT or look it up on
Google! At this point, I have code that doesn't work as I want it to,
but I have a definite update that I can make that will make it work or
at least improve what it is doing. Updated NOTES.md with thoughts.
Updated tsconfig so I could get access to prototype.toSpliced() ES2023.