feat: add function to get all edge cells
Added getAllEdgeCells to utils. Takes a grid as param and returns an
array of indices where an index is the index of a void at the edge of
the crossword grid. Updated ISSUES.md with notes about my problems
creating arrays of contiguous voids. Added notes to the mergeSubarrays
function.