feat: add more code for grid validation
Added isSubset function to utils. Returns true if all of the void
indices in an array are also in an array of the indices of the cells at
the edge of the grid. If all voids are on the edge, then they can't
cause islands of clues, and so we can set the isValid prop on the
corresponding cells in our grid to be true. Added code to the Grid.tsx
handleClick function to achieve the above.