feat: indicate island of answers with bg color
Renamed the data folder to 'state.' Added backgroundColors.ts to the
state folder. This is an array that holds hex colors I can use to set
the background colors for groups of cells that are islands. Added
optional backgroundColor prop to CellType. If backGround color is set,
use that for the color, else if the cell isVoid use black, otherwise use
white. Added getContiguousLights to utils. Moved mergeSubarrays to
utils from Grid. Moved code to check for groups of disconnected cells
inside the definition of validateGrid in Grid.tsx. At this point, if the
user toggles a cell's isVoid prop, we go and check if there are islands
of disconnected cells in the grid and set the background color for each
group.