feat: add isValid state to control button enabled
Added isValid and setIsValid to our Grid state. Our validdateGrid is
called in handleClick, so the grid is validated each time the user
toggles a cell from light to dark and vice-versa. The isValid state is
true by default and is set to true at the top of validateGrid each time
it is called. Once we validate the grid, we set isValid to false if
appropriate and this enables/disables buttons in the UI. If the grid is
invalid, the user should not be able to generate the answers, or the
clues, or save the crossword.