feat: an edge of all voids is highlighted to user
Added setAllVoidEdgeInvalid, which checks to see if any edge of the
crossword grid consists entirely of voids, and if it does it sets the
isValid prop for all those cells to false. This sets the border of each
cell to red, and so we can let the user know that the grid is not in a
valid state.