feat: add func to set isValid on initialization
Added isGridValid in utils. This checks whether the backgroundColor, and
isValid prop on each cell in the grid to determine if the grid is in a
valid state. Returns true if the grid is in a valid state, or false
otherwise. In Grid.tsx, we call this in the useState hook for the
isValid state if there is local storage. If there is no local storage,
then we are starting with the basic grid and this is valid and so we
return true. The user can now navigate away from the Create/Edit route,
navigate back, and the isValid state will be initialized correctly, ie.
buttons will be appropriately disabled.