feat: more code for tutorial/walkthrough
Added ArrowLeft.tsx which takes props top and left for positioning.
Added useRef to Grid.tsx so we can ref our button, the thing that we
want our ArrowLeft to point towards. Added position state to Grid which
is set with a useEffect if the ref exists. This position and useEffect
should reside in our ArrowLeft and we pass the ref to it. Changed name
of app to GridMaster - one of the options ChatGPT suggested!