refactor: update Grid.tsx to use columns
My ArrowLeft now points towards the button in my ref.current. Refactored
Grid.tsx so it consists of 3 columns: 1 for the control buttons, 1 for
the grid, and 1 to keep the spacing so the grid is more centered. Since
ArrowLeft is absolutely positioned and the button we are trying to point
at is not absolute, we have to subtract 48px (16 x 3 or 3 rem, the
height of the nav) from its top position. We then have to translate the
Y coord of our Arrow minus 25%, since we want the center of our Arrow
to point to the center of our button, rather than the top.